extract_load_scripts/requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
et-xmlfile==1.1.0
google-api-core==2.1.0
google-auth==2.3.0
google-cloud-bigquery==2.28.1
google-cloud-core==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
grpcio==1.41.0
idna==3.3
lxml==4.6.3
multitasking==0.0.9
numpy==1.21.2
openpyxl==3.0.9
packaging==21.0
pandas==1.3.3
proto-plus==1.19.5
protobuf==3.12.2
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
rsa==4.7.2
six==1.16.0
urllib3==1.26.7
yfinance==0.1.63